+++ src/core/globs.ts
@@ -1,0 +1,5 @@
+export const GLOB_CSS = '**/*.css';
+export const GLOB_JSON = '**/*.json';
+export const GLOB_JSONC = '**/*.jsonc';
+export const GLOB_SRC = '**/*.?([cm])[jt]s?(x)';
+export const GLOB_VUE = '**/*.vue';